requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
ldap3==2.9.1
MarkupSafe==3.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-ldap==3.4.4
Werkzeug==3.1.3
